Fix imports for sqlite migrator
Fixes issue when sqlite is being used in non node environments, where importing from migrator.ts causes issues because of the node:crypto, node:fs and node:path.
