+++ drizzle-orm/src/durable-sqlite/migrator.ts
@@ -1,4 +1,4 @@
-import { formatToMillis, type MigrationMeta, type MigratorInitFailResponse } from '~/migrator.ts';
+import { formatToMillis, type MigrationMeta, type MigratorInitFailResponse } from '~/migrator-utils.ts';
 import type { AnyRelations } from '~/relations.ts';
 import { sql } from '~/sql/index.ts';
 import type { DrizzleSqliteDODatabase } from './driver.ts';
