+++ drizzle-orm/src/op-sqlite/migrator.ts
@@ -1,5 +1,5 @@
 import { useEffect, useReducer } from 'react';
-import { formatToMillis, type MigrationMeta } from '~/migrator.ts';
+import { formatToMillis, type MigrationMeta } from '~/migrator-utils.ts';
 import type { AnyRelations } from '~/relations.ts';
 import type { OPSQLiteDatabase } from './driver.ts';
